update country context to settings context
